fix onda 4: crate::history no poller, Debug no Payload, casts usize/u32 do arboard, keyring so no Windows (runner Linux sem dbus-1)

+++ apps/lab-clip/src/main.rs
@@ -142,8 +142,8 @@ impl ClipApp {
         };
         let rgba = dec.to_rgba8();
         let data = arboard::ImageData {
-            width: rgba.width(),
-            height: rgba.height(),
+            width: rgba.width() as usize,
+            height: rgba.height() as usize,
             bytes: std::borrow::Cow::Owned(rgba.into_raw()),
         };
         if let Ok(mut c) = arboard::Clipboard::new() {
